requirements.txt
python-dotenv==0.20.0
gunicorn==20.1.0
Flask==2.0.2
Flask-Cors==3.0.10
backoff==2.2.1
PyYAML==6.0
pytest==7.4.0
openai==0.27.8
tiktoken==0.2.0
pytz==2023.3